moves OrnamentDictionary into own class for sololy use
