+++ src/meico/mei/ornament/OrnamentExpander.java
@@ -36,26 +36,10 @@ public class OrnamentExpander {
     private Map<String, MeiElementHelper> currentNotes                = new HashMap<>();   // cached notes in the current measure, to be used for grace note expansion
 
     /**
-     * constructor
-     * @param mei the MEI to be expanded
-     */
-    public OrnamentExpander(Mei mei) {
-        try {
-            createOrnamentLookUp();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
      * default constructor, if MEI is not yet available
      */
     public OrnamentExpander()  {
-        try {
-            createOrnamentLookUp();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        this.ornamentLookup = new OrnamentDictionary().getOrnamentLookup();
     }
 
     /**
@@ -706,61 +690,5 @@ public class OrnamentExpander {
             nextOrnams.remove(ornament.get("next")); // remove myself if I was in there
             expandOrnamentsElement(nextOrnam);
         }
-    }
-
-    /**
-     * creates the lookUp table with ornament descriptions. Names for lookUp are identical to getOrnamentFullName results
-     */
-    private void createOrnamentLookUp() throws IOException, NullPointerException {
-        this.ornamentLookup = new HashMap<String, List<String>>();
-
-        // open input stream
-        InputStream is = getClass().getResourceAsStream("/resources/ornaments.dict");
-        if(is == null)
-            return;
-
-        // initialize the readers with the input stream
-        InputStreamReader ir = new InputStreamReader(is);
-        BufferedReader br = new BufferedReader(ir);
-
-        // build (key, value) pairs where the key is the ornament name string and value is the List of alterations and add them to the dict map
-        List<String> ornamentNames = new ArrayList<String>();
-        boolean isCollectingNames = false;
-        for(String line = br.readLine(); line != null; line = br.readLine()) {  // read all the lines in *.dict
-            if (line.isEmpty()                                                  // an empty line
-                    || (line.charAt(0) == '%'))                                     // this is a comment line
-                continue;                                                       // ignore it
-
-            if (line.charAt(0) == '#') {                                        // this is an ornament name line, it specifies that all further lines will be associated with it until an ornament line is read
-                if(!isCollectingNames) {                                        // if it is the first line with an ornament name
-                    isCollectingNames = true;
-                    ornamentNames.clear();
-                }
-                ornamentNames.add(line.substring(1).trim());                 // add the ornamentName, delete any spaces in the string beforehand so that "# trill " -> "trill"
-
-                continue;
-            }
-            else if (isCollectingNames) {                                       // if I am currently collecting names, but I have read a line that is not an ornament name line, I am now collecting alterations for the current ornament name
-                isCollectingNames = false;
-            }
-
-            if(!line.isEmpty()) {
-                ArrayList<String> alterationEntries = new ArrayList<String>(Arrays.asList(line.split(" "))); // split the line by " " to get the single alteration entries, e.g. "1 0 -1" -> ["1", "0", "-1"]
-
-                for (String ornamentName : ornamentNames) {                      // for all currently collected ornament names, add the alteration to the lookUp table
-                    ornamentLookup.put(ornamentName, new ArrayList<String>());
-                    List<String> lookUp = ornamentLookup.get(ornamentName);
-                    for(String alterationEntry : alterationEntries) {            // cleanUp the alteration entries, e.g. ["", "1 "] -> ["1"]
-                        if(!alterationEntry.isEmpty())
-                            lookUp.add(alterationEntry.trim());
-                    }
-                }
-            }
-        }
-
-        // close readers and input stream
-        br.close();
-        ir.close();
-        is.close();
     }
 }
